+++ components/camel-pulsar/src/test/java/org/apache/camel/component/pulsar/integration/PulsarConcurrentProducerInIT.java
@@ -14,8 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.pulsar;
+package org.apache.camel.component.pulsar.integration;
 
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 
 import org.apache.camel.Endpoint;
@@ -24,6 +26,7 @@ import org.apache.camel.Produce;
 import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
+import org.apache.camel.component.pulsar.PulsarComponent;
 import org.apache.camel.component.pulsar.utils.AutoConfiguration;
 import org.apache.camel.spi.Registry;
 import org.apache.camel.support.SimpleRegistry;
@@ -32,26 +35,18 @@ import org.apache.pulsar.client.api.PulsarClientException;
 import org.apache.pulsar.client.impl.ClientBuilderImpl;
 import org.junit.jupiter.api.Test;
 
-public class PulsarProducerUndefinedProducerNameInTest extends PulsarTestSupport {
+public class PulsarConcurrentProducerInIT extends PulsarITSupport {
 
-    private static final String TOPIC_URI = "persistent://public/default/camel-producer-topic";
+    private static final String TOPIC_URI = "persistent://public/default/camel-concurrent-producers-topic";
+    private static final String PRODUCER = "camel-producer";
 
-    @Produce("direct:start1")
-    private ProducerTemplate producerTemplate1;
+    @Produce("direct:start")
+    private ProducerTemplate producerTemplate;
 
-    @Produce("direct:start2")
-    private ProducerTemplate producerTemplate2;
-
-    @EndpointInject("pulsar:" + TOPIC_URI
-                    + "?numberOfConsumers=1"
-                    + "&subscriptionType=Exclusive"
-                    + "&subscriptionName=camel-subscription"
-                    + "&consumerQueueSize=1"
-                    + "&consumerName=camel-consumer")
-    private Endpoint pulsarEndpoint1;
-
-    @EndpointInject("pulsar:" + TOPIC_URI)
-    private Endpoint pulsarEndpoint2;
+    @EndpointInject("pulsar:" + TOPIC_URI + "?numberOfConsumers=3&subscriptionType=Shared"
+                    + "&subscriptionName=camel-subscription&consumerQueueSize=5"
+                    + "&consumerNamePrefix=camel-consumer" + "&producerName=" + PRODUCER)
+    private Endpoint from;
 
     @EndpointInject("mock:result")
     private MockEndpoint to;
@@ -62,24 +57,22 @@ public class PulsarProducerUndefinedProducerNameInTest extends PulsarTestSupport
 
             @Override
             public void configure() {
-                from("direct:start1").to(pulsarEndpoint1);
-                from("direct:start2").to(pulsarEndpoint2);
-
-                from(pulsarEndpoint1).to(to);
+                from("direct:start").to(from);
+                from(from).to(to);
             }
         };
     }
 
     @Override
     protected Registry createCamelRegistry() throws Exception {
-        Registry registry = new SimpleRegistry();
+        SimpleRegistry registry = new SimpleRegistry();
 
         registerPulsarBeans(registry);
 
         return registry;
     }
 
-    private void registerPulsarBeans(Registry registry) throws PulsarClientException {
+    private void registerPulsarBeans(SimpleRegistry registry) throws PulsarClientException {
         PulsarClient pulsarClient = givenPulsarClient();
         AutoConfiguration autoConfiguration = new AutoConfiguration(null, null);
 
@@ -88,23 +81,34 @@ public class PulsarProducerUndefinedProducerNameInTest extends PulsarTestSupport
         comp.setAutoConfiguration(autoConfiguration);
         comp.setPulsarClient(pulsarClient);
         registry.bind("pulsar", comp);
+
     }
 
     private PulsarClient givenPulsarClient() throws PulsarClientException {
-        return new ClientBuilderImpl()
-                .serviceUrl(getPulsarBrokerUrl())
-                .ioThreads(1)
-                .listenerThreads(1)
-                .build();
+        return new ClientBuilderImpl().serviceUrl(getPulsarBrokerUrl()).ioThreads(1).listenerThreads(1).build();
     }
 
     @Test
-    public void testAMessageToRouteIsSentFromBothProducersAndThenConsumed() throws Exception {
-        to.expectedMessageCount(2);
+    public void testConcurrentMessagesAreSentAndThenConsumed() throws Exception {
+        to.expectedMessageCount(100);
 
-        producerTemplate1.sendBody("Test First");
-        producerTemplate2.sendBody("Test Second");
+        sendMessages();
 
         MockEndpoint.assertIsSatisfied(10, TimeUnit.SECONDS, to);
+    }
+
+    private void sendMessages() {
+        ExecutorService executorService = Executors.newFixedThreadPool(5);
+
+        for (int i = 0; i < 100; i++) {
+            executorService.submit(new Runnable() {
+                @Override
+                public void run() {
+                    producerTemplate.sendBody("Hello World!");
+                }
+            });
+        }
+
+        executorService.shutdown();
     }
 }
