+++ components/camel-pulsar/src/test/java/org/apache/camel/component/pulsar/integration/PulsarConsumerReadCompactedIT.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.pulsar;
+package org.apache.camel.component.pulsar.integration;
 
 import java.util.concurrent.TimeUnit;
 
@@ -23,6 +23,8 @@ import org.apache.camel.Endpoint;
 import org.apache.camel.EndpointInject;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
+import org.apache.camel.component.pulsar.PulsarComponent;
+import org.apache.camel.component.pulsar.PulsarMessageReceipt;
 import org.apache.camel.component.pulsar.utils.AutoConfiguration;
 import org.apache.camel.component.pulsar.utils.message.PulsarMessageHeaders;
 import org.apache.camel.spi.Registry;
@@ -44,8 +46,8 @@ import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class PulsarConsumerReadCompactedTest extends PulsarTestSupport {
-    private static final Logger LOGGER = LoggerFactory.getLogger(PulsarConsumerReadCompactedTest.class);
+public class PulsarConsumerReadCompactedIT extends PulsarITSupport {
+    private static final Logger LOGGER = LoggerFactory.getLogger(PulsarConsumerReadCompactedIT.class);
 
     private static final String TOPIC_URI = "persistent://public/default/camel-topic";
