+++ components/camel-pulsar/src/test/java/org/apache/camel/component/pulsar/integration/PulsarProducerInIT.java
@@ -14,10 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.pulsar;
+package org.apache.camel.component.pulsar.integration;
 
-import java.util.HashMap;
-import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
 import org.apache.camel.Endpoint;
@@ -26,8 +24,8 @@ import org.apache.camel.Produce;
 import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
+import org.apache.camel.component.pulsar.PulsarComponent;
 import org.apache.camel.component.pulsar.utils.AutoConfiguration;
-import org.apache.camel.component.pulsar.utils.message.PulsarMessageHeaders;
 import org.apache.camel.spi.Registry;
 import org.apache.camel.support.SimpleRegistry;
 import org.apache.pulsar.client.api.PulsarClient;
@@ -35,7 +33,7 @@ import org.apache.pulsar.client.api.PulsarClientException;
 import org.apache.pulsar.client.impl.ClientBuilderImpl;
 import org.junit.jupiter.api.Test;
 
-public class PulsarProducerHeadersInTest extends PulsarTestSupport {
+public class PulsarProducerInIT extends PulsarITSupport {
 
     private static final String TOPIC_URI = "persistent://public/default/camel-producer-topic";
     private static final String PRODUCER = "camel-producer";
@@ -43,16 +41,13 @@ public class PulsarProducerHeadersInTest extends PulsarTestSupport {
     @Produce("direct:start")
     private ProducerTemplate producerTemplate;
 
-    @EndpointInject("pulsar:" + TOPIC_URI
-                    + "?numberOfConsumers=1&subscriptionType=Exclusive"
-                    + "&subscriptionName=camel-subscription"
-                    + "&consumerQueueSize=1"
-                    + "&consumerName=camel-consumer"
-                    + "&producerName=" + PRODUCER)
-    private Endpoint pulsar;
+    @EndpointInject("pulsar:" + TOPIC_URI + "?numberOfConsumers=1&subscriptionType=Exclusive"
+                    + "&subscriptionName=camel-subscription&consumerQueueSize=1"
+                    + "&consumerName=camel-consumer" + "&producerName=" + PRODUCER)
+    private Endpoint from;
 
     @EndpointInject("mock:result")
-    private MockEndpoint mock;
+    private MockEndpoint to;
 
     @Override
     protected RouteBuilder createRouteBuilder() {
@@ -60,22 +55,22 @@ public class PulsarProducerHeadersInTest extends PulsarTestSupport {
 
             @Override
             public void configure() {
-                from("direct:start").to(pulsar);
-                from(pulsar).to(mock);
+                from("direct:start").to(from);
+                from(from).to(to);
             }
         };
     }
 
     @Override
     protected Registry createCamelRegistry() throws Exception {
-        Registry registry = new SimpleRegistry();
+        SimpleRegistry registry = new SimpleRegistry();
 
         registerPulsarBeans(registry);
 
         return registry;
     }
 
-    private void registerPulsarBeans(final Registry registry) throws PulsarClientException {
+    private void registerPulsarBeans(SimpleRegistry registry) throws PulsarClientException {
         PulsarClient pulsarClient = givenPulsarClient();
         AutoConfiguration autoConfiguration = new AutoConfiguration(null, null);
 
@@ -84,44 +79,21 @@ public class PulsarProducerHeadersInTest extends PulsarTestSupport {
         comp.setAutoConfiguration(autoConfiguration);
         comp.setPulsarClient(pulsarClient);
         registry.bind("pulsar", comp);
+
     }
 
     private PulsarClient givenPulsarClient() throws PulsarClientException {
-        return new ClientBuilderImpl()
-                .serviceUrl(getPulsarBrokerUrl())
-                .ioThreads(1)
-                .listenerThreads(1)
-                .build();
+        return new ClientBuilderImpl().serviceUrl(getPulsarBrokerUrl()).ioThreads(1).listenerThreads(1).build();
     }
 
     @Test
-    public void propertyHeaderSetsPulsarProperties() throws InterruptedException {
-        Map<String, String> properties = new HashMap<>();
-        properties.put("testProperty", "testValue");
-        mock.expectedHeaderReceived(PulsarMessageHeaders.PROPERTIES, properties);
+    public void testAMessageToRouteIsSentAndThenConsumed() throws Exception {
+        to.expectedMessageCount(3);
 
-        producerTemplate.sendBodyAndHeader("test", PulsarMessageHeaders.PROPERTIES_OUT, properties);
+        producerTemplate.sendBody("Hello ");
+        producerTemplate.sendBody("World ");
+        producerTemplate.sendBody(10);
 
-        MockEndpoint.assertIsSatisfied(10, TimeUnit.SECONDS, mock);
-    }
-
-    @Test
-    public void eventTimeHeaderSetsPulsarEventTime() throws InterruptedException {
-        long eventTime = 10000;
-        mock.expectedHeaderReceived(PulsarMessageHeaders.EVENT_TIME, eventTime);
-
-        producerTemplate.sendBodyAndHeader("test", PulsarMessageHeaders.EVENT_TIME_OUT, eventTime);
-
-        MockEndpoint.assertIsSatisfied(10, TimeUnit.SECONDS, mock);
-    }
-
-    @Test
-    public void keyHeaderSetsPulsarKey() throws InterruptedException {
-        String key = "testKey";
-        mock.expectedHeaderReceived(PulsarMessageHeaders.KEY, key);
-
-        producerTemplate.sendBodyAndHeader("test", PulsarMessageHeaders.KEY_OUT, key);
-
-        MockEndpoint.assertIsSatisfied(10, TimeUnit.SECONDS, mock);
+        MockEndpoint.assertIsSatisfied(10, TimeUnit.SECONDS, to);
     }
 }
